topic, comment, comment historyのポリシーのマルチテナント対応
