--- app/policies/comment_histories_policy.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-class CommentHistoriesPolicy < ApplicationPolicy
-  def index? = owner? || moderator?
-  def compare? = owner? || moderator?
-end
