--- app/policies/comments_policy.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class CommentsPolicy < ApplicationPolicy
-  def create? = unsuspended_user?
-  def edit? = unsuspended_user? && owner?
-  def update? = unsuspended_user? && owner?
-end
