+++ app/policies/tenants/comment_histories_policy.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Tenants
+  class CommentHistoriesPolicy < ApplicationPolicy
+    def index? = (tenant_member? && owner?) || moderator?
+    def compare? = (tenant_member? && owner?) || moderator?
+  end
+end
