+++ app/policies/tenants/topics_policy.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Tenants
+  class TopicsPolicy < ApplicationPolicy
+    def index? = true
+    def show? = true
+    def new? = tenant_member? && unsuspended_user?
+    def create? = tenant_member? && unsuspended_user?
+    def edit? = tenant_member? && unsuspended_user? && owner?
+    def update? = tenant_member? && unsuspended_user? && owner?
+  end
+end
